external_repos/pluginSHIRE/containers/v6-http-test/requirements.txt
vantage6-algorithm-tools
pandas
aiohttp
requests
fsspec